move remaining definitions
